+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidyAML (development version)
+
 # tidyAML 0.0.4
 
 ## Breaking Changes