+++ NEWS.md
@@ -1,4 +1,4 @@
-# tidyAML (development version)
+# tidyAML 0.0.5
 
 ## Breaking Changes
 None